requirements.txt
annotated-types==0.7.0
anyio==4.4.0
barcode==1.0.4
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
docker==7.1.0
fastapi==0.112.1
gunicorn==23.0.0
h11==0.14.0
idna==3.7
Levenshtein==0.25.1
loguru==0.7.2
lxml==5.3.0
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pypdfium2==4.30.0
pypng==0.20220715.0
pytesseract==0.3.13
python-barcode==0.15.1
python-docx==1.1.2
python-Levenshtein==0.25.1
python-multipart==0.0.9
qrcode==7.4.2
rapidfuzz==3.9.7
requests==2.32.3
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
